Another update
Added:
+ Update checker (Can be disabled in the config)
+ An option to overwrite saves (Disabled by default)
If you are using MySQL and the database detects that a recording is already saved, (If you kick players instead of banning and they rejoin and instantly hack again) an error will be printed in console. With this disabled a random number between 1 - 100 will be added to the end of the file name so you can save different recordings of a player hacking.

Fixed:
Attempted to make the plugin compatible with PlugMan or other reloading plugins. This is probably the most I will do to attempt to support these types of plugins. I still recommend rebooting the server when you install/edit my plugin.